move shallow render main test from game to welcome
